Move types-v3/utils-v3 into v3 folder, restore test_mode setting
- Move types-v3.ts and utils-v3.ts into a new v3/ subfolder, fixing
  their relative imports and all call sites
- Re-add the test_mode Settings field (removed earlier, but still
  referenced by v2 utils.ts) and regenerate generated-types.ts +
  metadata.json
- Update test_mode's description to clarify it's V2-only (deprecated)
  and point to the per-action Test ID field for V3 test events
