+++ packages/destination-actions/src/destinations/reddit-conversions-api/index.ts
@@ -25,6 +25,14 @@ const destination: DestinationDefinition<Settings> = {
           'The conversion token for your Reddit account. This can be found by following the steps mentioned [here](https://business.reddithelp.com/helpcenter/s/article/conversion-access-token).',
         type: 'password',
         required: true
+      },
+      test_mode: {
+        label: '[Deprecated] Test Mode',
+        description:
+          'Indicates if events should be treated as test events by Reddit. Only applies to Reddit Conversions API V2, which is deprecated - V3 is the latest API version. To send test events on V3, set the Test ID field on an action mapping instead.',
+        type: 'boolean',
+        required: false,
+        default: false
       }
     },
     testAuthentication: async (request, { settings }) => {
