+++ packages/destination-actions/src/destinations/reddit-conversions-api/metadata.json
@@ -27,8 +27,8 @@
         "depends_on": null
       },
       "test_mode": {
-        "label": "Test Mode",
-        "description": "Indicates if events should be treated as test events by Reddit.",
+        "label": "[Deprecated] Test Mode",
+        "description": "Indicates if events should be treated as test events by Reddit. Only applies to Reddit Conversions API V2, which is deprecated - V3 is the latest API version. To send test events on V3, set the Test ID field on an action mapping instead.",
         "type": "boolean",
         "required": false,
         "multiple": false,
